Collision update

+++ src/Tickable.java
@@ -1,4 +1,4 @@
 public abstract class Tickable {
     
-    public abstract void tick(float deltaTime);
+    public abstract void tick(double deltaTime);
 }
